+++ src/main/java/nextstep/subway/line/dto/LineResponse.java
@@ -6,7 +6,6 @@ import nextstep.subway.station.dto.StationResponse;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
 
 public class LineResponse {
     private Long id;
