+++ src/test/java/nextstep/subway/line/acceptance/LineAcceptanceTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line;
+package nextstep.subway.line.acceptance;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
