+++ src/test/java/nextstep/subway/line/acceptance/SectionAcceptanceTest.java
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.test.util.ReflectionTestUtils;
 
 import static nextstep.subway.line.acceptance.LineSteps.*;
 import static nextstep.subway.line.acceptance.SectionSteps.*;
