+++ src/test/java/nextstep/subway/line/application/LineServiceMockTest.java
@@ -1,13 +1,10 @@
 package nextstep.subway.line.application;
 
-import io.restassured.internal.RestAssuredResponseOptionsImpl;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
-import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.line.dto.SectionRequest;
 import nextstep.subway.station.application.StationService;
 import nextstep.subway.station.domain.Station;
-import nextstep.subway.station.dto.StationRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
